+++ matchbox-server/src/main/java/ch/ahdis/matchbox/gazelle/GazelleValidationWs.java
@@ -27,6 +27,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
+import static ch.ahdis.matchbox.util.MatchboxPackageInstallerImpl.SD_EXTENSION_TITLE_PREFIX;
+
 /**
  * The WebService for validation with the new Gazelle Validation API.
  *
@@ -93,7 +95,9 @@ public class GazelleValidationWs {
 	 */
 	@GetMapping(path = PROFILES_PATH, produces = MediaType.APPLICATION_JSON_VALUE)
 	public List<ValidationProfile> getProfiles() {
+		// Filter the extensions, because they won't be validated directly
 		return this.structureDefinitionProvider.getPackageResources().stream()
+			.filter(packageVersionResource -> !packageVersionResource.getFilename().startsWith(SD_EXTENSION_TITLE_PREFIX))
 			.map(packageVersionResource -> {
 				final var profile = new ValidationProfile();
 				final var version = packageVersionResource.getCanonicalVersion();